1/11

+++ management/src/components/topNav.vue
@@ -139,10 +139,29 @@ export default {
     },
     //点击注销
     cancellationClick() {
-      sessionStorage.setItem('toke',1)
-        this.$router.push({
+      
+        this.$confirm('确定退出登陆吗?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+              sessionStorage.setItem('toke',1)
+               this.$router.push({
           path:'/'
         })
+          this.$message({
+            type: 'success',
+            message: '退出成功!'
+          });
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消退出'
+          });          
+        });
+      
+  
+       
     }
   },
